soc/mediatek/mt8186: Reduce GPIO code size in bootblock
Create a new GPIO driving info table that contains only the pins used
in the bootblock. The GPIO driving info table is downsized from 1480
bytes to 24 bytes.

BUG=b:270911452
TEST=build pass

Reviewed-on: https://review.coreboot.org/c/coreboot/+/74062
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/mediatek/common/include/soc/gpio_common.h
@@ -34,7 +34,7 @@ enum gpio_drv_adv {
 };
 
 struct gpio_drv_info {
-	uint32_t offset;
+	uint8_t offset;
 	uint8_t shift;
 	uint8_t width;
 };
